requirements.txt
streamlit
requests
pandas
numpy
scikit-learn
textblob
python-dotenv
plotly